test_deep_subdomain_integration.py
"""
Test script to verify Deep Subdomain Scanner integration
Ensures:
1. No breaking changes to existing functionality
2. New deep scan mode works correctly
3. All imports are available
4. Output format is correct
"""

import sys
import os

# Add src to path
sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))

def test_imports():
    """Test that all required modules import successfully"""
    print("=" * 60)
    print("Testing Imports...")
    print("=" * 60)
    
    try:
        from core.scanner import check_subdomain
        print("✅ core.scanner.check_subdomain imported successfully")
    except Exception as e:
        print(f"❌ Failed to import check_subdomain: {e}")
        return False
    
    try:
        from core.deep_subdomain_scanner import DeepSubdomainScanner
        print("✅ core.deep_subdomain_scanner.DeepSubdomainScanner imported successfully")
    except Exception as e:
        print(f"❌ Failed to import DeepSubdomainScanner: {e}")
        return False
    
    try:
        from core.deep_subdomain_scanner import scan_subdomains_blocking
        print("✅ core.deep_subdomain_scanner.scan_subdomains_blocking imported successfully")
    except Exception as e:
        print(f"❌ Failed to import scan_subdomains_blocking: {e}")
        return False
    
    try:
        import dns.resolver
        print("✅ dns.resolver imported successfully")
    except Exception as e:
        print(f"❌ Failed to import dns.resolver: {e}")
        return False
    
    return True


def test_existing_functionality():
    """Test that existing check_subdomain function still works"""
    print("\n" + "=" * 60)
    print("Testing Existing Functionality (Backward Compatibility)...")
    print("=" * 60)
    
    try:
        from core.scanner import check_subdomain
        
        # Test with a well-known domain
        # Note: This may fail in isolated environments without DNS
        result = check_subdomain("google.com", "www")
        print(f"✅ check_subdomain('google.com', 'www') executed (result: {result})")
        
        # Test with invalid subdomain (should return False)
        result = check_subdomain("invalid-test-domain-12345.com", "xyz")
        print(f"✅ check_subdomain returned False for invalid domain (result: {result})")
        
        return True
    except Exception as e:
        print(f"⚠️  Existing functionality test skipped (DNS may not be available): {e}")
        return True  # Not a failure, just unavailable in test environment


def test_scanner_initialization():
    """Test that DeepSubdomainScanner initializes correctly"""
    print("\n" + "=" * 60)
    print("Testing DeepSubdomainScanner Initialization...")
    print("=" * 60)
    
    try:
        from core.deep_subdomain_scanner import DeepSubdomainScanner
        
        # Test simple mode
        scanner_simple = DeepSubdomainScanner("example.com", deep_scan=False)
        print(f"✅ DeepSubdomainScanner initialized in simple mode")
        print(f"   - Domain: {scanner_simple.domain}")
        print(f"   - Deep Scan: {scanner_simple.deep_scan}")
        print(f"   - Max Workers: {scanner_simple.max_workers}")
        
        # Test deep mode
        scanner_deep = DeepSubdomainScanner("example.com", deep_scan=True)
        print(f"✅ DeepSubdomainScanner initialized in deep mode")
        print(f"   - Domain: {scanner_deep.domain}")
        print(f"   - Deep Scan: {scanner_deep.deep_scan}")
        print(f"   - Max Workers: {scanner_deep.max_workers}")
        
        return True
    except Exception as e:
        print(f"❌ Failed to initialize scanner: {e}")
        import traceback
        traceback.print_exc()
        return False


def test_wordlist_loading():
    """Test that wordlist loads correctly"""
    print("\n" + "=" * 60)
    print("Testing Wordlist Loading...")
    print("=" * 60)
    
    try:
        from core.deep_subdomain_scanner import DeepSubdomainScanner
        
        scanner = DeepSubdomainScanner("example.com", deep_scan=True)
        wordlist = scanner.load_wordlist()
        
        print(f"✅ Wordlist loaded successfully")
        print(f"   - Total entries: {len(wordlist)}")
        print(f"   - Sample entries: {wordlist[:5]}")
        
        if len(wordlist) > 0:
            print(f"✅ Wordlist contains {len(wordlist)} subdomains")
            return True
        else:
            print(f"❌ Wordlist is empty")
            return False
            
    except Exception as e:
        print(f"❌ Failed to load wordlist: {e}")
        import traceback
        traceback.print_exc()
        return False


def test_permutation_generation():
    """Test that subdomain permutations are generated"""
    print("\n" + "=" * 60)
    print("Testing Permutation Generation...")
    print("=" * 60)
    
    try:
        from core.deep_subdomain_scanner import DeepSubdomainScanner
        
        scanner = DeepSubdomainScanner("example.com", deep_scan=True)
        base_subs = ["api", "admin", "dev", "test"]
        perms = scanner.generate_permutations(base_subs)
        
        print(f"✅ Permutations generated successfully")
        print(f"   - Base subdomains: {len(base_subs)}")
        print(f"   - Total permutations: {len(perms)}")
        print(f"   - Sample permutations: {perms[:10]}")
        
        if len(perms) > len(base_subs):
            print(f"✅ Permutation logic working correctly")
            return True
        else:
            print(f"⚠️  Permutations generated but may be fewer than expected")
            return True
            
    except Exception as e:
        print(f"❌ Failed to generate permutations: {e}")
        import traceback
        traceback.print_exc()
        return False


def test_output_format():
    """Test that output format is correct"""
    print("\n" + "=" * 60)
    print("Testing Output Format...")
    print("=" * 60)
    
    try:
        from core.deep_subdomain_scanner import SubdomainResult
        
        # Test SubdomainResult creation
        result = SubdomainResult(
            subdomain="api.example.com",
            status_code=200,
            status_text="Live",
            dns_records={"A": ["192.168.1.1"]}
        )
        
        result_dict = result.to_dict()
        
        print(f"✅ SubdomainResult object created and converted to dict")
        print(f"   - Subdomain: {result_dict['subdomain']}")
        print(f"   - Status Code: {result_dict['status_code']}")
        print(f"   - Status Text: {result_dict['status_text']}")
        print(f"   - DNS Records: {result_dict['dns_records']}")
        
        # Verify all required fields
        required_fields = ['subdomain', 'status_code', 'status_text', 'dns_records']
        if all(field in result_dict for field in required_fields):
            print(f"✅ Output format contains all required fields")
            return True
        else:
            print(f"❌ Output format missing required fields")
            return False
            
    except Exception as e:
        print(f"❌ Failed to test output format: {e}")
        import traceback
        traceback.print_exc()
        return False


def main():
    """Run all tests"""
    print("\n")
    print("╔" + "=" * 58 + "╗")
    print("║" + " " * 58 + "║")
    print("║" + "  Deep Subdomain Scanner - Integration Test Suite  ".center(58) + "║")
    print("║" + " " * 58 + "║")
    print("╚" + "=" * 58 + "╝")
    
    tests = [
        ("Imports", test_imports),
        ("Existing Functionality", test_existing_functionality),
        ("Scanner Initialization", test_scanner_initialization),
        ("Wordlist Loading", test_wordlist_loading),
        ("Permutation Generation", test_permutation_generation),
        ("Output Format", test_output_format),
    ]
    
    results = []
    for test_name, test_func in tests:
        try:
            result = test_func()
            results.append((test_name, result))
        except Exception as e:
            print(f"\n❌ Test '{test_name}' failed with exception: {e}")
            import traceback
            traceback.print_exc()
            results.append((test_name, False))
    
    # Summary
    print("\n" + "=" * 60)
    print("Test Summary")
    print("=" * 60)
    
    passed = sum(1 for _, result in results if result)
    total = len(results)
    
    for test_name, result in results:
        status = "✅ PASS" if result else "❌ FAIL"
        print(f"{status}: {test_name}")
    
    print(f"\nTotal: {passed}/{total} tests passed")
    
    if passed == total:
        print("\n🎉 All tests passed! Integration is successful!")
        return 0
    else:
        print(f"\n❌ {total - passed} test(s) failed. Please review the errors above.")
        return 1


if __name__ == "__main__":
    sys.exit(main())
